Store position length alongside position in index. This is very much WIP as I haven't tested very many cases yet.
Modify exact phrase matcher to respect position length.

Sloppy phrase matching WIP.

+++ lucene/codecs/src/java/org/apache/lucene/codecs/simpletext/SimpleTextFieldsWriter.java
@@ -152,7 +152,7 @@ class SimpleTextFieldsWriter extends FieldsConsumer {
 
               // for each pos in field+term+doc
               for (int i = 0; i < freq; i++) {
-                int position = postingsEnum.nextPosition();
+                int position = (int) postingsEnum.nextPosition();
 
                 write(POS);
                 write(Integer.toString(position));
