+++ lucene/sandbox/src/java/org/apache/lucene/sandbox/codecs/idversion/SingleDocsEnum.java
@@ -68,7 +68,7 @@ class SingleDocsEnum extends PostingsEnum {
   }
 
   @Override
-  public int nextPosition() throws IOException {
+  public long nextPosition() throws IOException {
     return -1;
   }
